api/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
bcrypt
passlib[bcrypt]
pymongo
# Add any other required packages
pydantic[email]
python-multipart